requirements.txt
discord.py
aitextgen
pytorch-lightning==1.7.7
